added model name logic for pipeline 3 and completed it for pipeline 2
It works now in the following for all pipelines:
If a model name is provided dynamically as varargin, it will always take this name (highest priority).
If you provide a hard coded model name in settings.model, it will take this name.
If you provide no settings.model, but only a functionhandle, in either settings or dynamically, it will take this name.

I think in that way, the old workings of the pipeline is not interfered, but it has the additional flexibility of defining dynamic model names.